+++ src/commands/BaseExporter.ts
@@ -1,14 +1,6 @@
 import * as vscode from 'vscode';
-import { IdentityNowResourceTreeItem, SourceTreeItem } from '../models/IdentityNowTreeItem';
 import { IdentityNowClient } from '../services/IdentityNowClient';
-import { delay } from '../utils';
-import * as fs from 'fs';
-import { withQuery } from '../utils/UriUtils';
 import { CSVWriter } from '../services/CSVWriter';
-import { Schema } from '../models/Schema';
-import { off } from 'process';
-import { PathProposer } from '../services/PathProposer';
-import { askFile } from '../utils/vsCodeHelpers';
 import { ensureFolderExists } from '../utils/fileutils';
 
 /**
@@ -16,60 +8,70 @@ import { ensureFolderExists } from '../utils/fileutils';
  */
 export abstract class BaseCSVExporter<T> {
 
+    readonly client: IdentityNowClient;
 
 
-
-    constructor(protected objectType: string) { }
+    constructor(protected objectType: string,
+        protected tenantId: string,
+        protected tenantName: string,
+        protected tenantDisplayName: string,
+        protected sourceId: string,
+        protected filePath: string
+    ) {
+        this.client = new IdentityNowClient(tenantId, tenantName);
+    }
 
     /**
      * Will display a progress bar for the export
      */
-    public async exportFileWithProgression(
-        tenantName: string,
-        tenantDisplayName: string,
-        tenantId: string,
-        filePath: string,
-        sourceId: string
-    ): Promise<void> {
+    public async exportFileWithProgression(): Promise<void> {
 
-        const client = new IdentityNowClient(tenantId, tenantName);
         await vscode.window.withProgress({
             location: vscode.ProgressLocation.Notification,
-            title: `Exporting ${this.objectType} from ${tenantName}...`,
+            title: `Exporting ${this.objectType} from ${this.tenantName}...`,
             cancellable: false
-        }, async (task, token) => await this.exportFile(client, tenantName, tenantDisplayName, sourceId, filePath, task, token))
-            .then(async () =>
-                await vscode.window.showInformationMessage(
-                    `Successfully exported ${this.objectType} from ${tenantName}`
-                ));
+        }, async (task, token) =>
+            await this.exportFile(task, token)
+        )
+            .then(async () =>{
+                vscode.window.showInformationMessage(
+                    `Successfully exported ${this.objectType} from ${this.tenantName}`
+                );
+                const document = await vscode.workspace.openTextDocument(vscode.Uri.file(this.filePath));
+                vscode.window.showTextDocument(document);
+            });
     }
 
+    /**
+     * 
+     * @param headers Headers of CSV
+     * @param paths Dotted path 
+     * @param unwindablePaths 
+     * @param iterator 
+     */
     protected async writeData(
-        filePath: string,
         headers: string[],
         paths: string[],
         unwindablePaths: string[],
-        iterator: AsyncIterable<T[]>) {
-
+        iterator: AsyncIterable<T[]>,
+        task: any, token: vscode.CancellationToken) {
+        ensureFolderExists(this.filePath);
         const csvWriter = new CSVWriter(
-            filePath,
+            this.filePath,
             headers,
             paths,
             unwindablePaths);
 
         for await (const data of iterator) {
             await csvWriter.write(data);
+            if (token.isCancellationRequested) {
+                break;
+            }
         }
-        csvWriter.end();
+        await csvWriter.end();
     }
 
-    protected abstract exportFile(
-        client: IdentityNowClient,
-        tenantName: string,
-        tenantDisplayName: string,
-        sourceId: string,
-        filePath: string,
-        task: any, token: vscode.CancellationToken): Promise<void>;
+    protected abstract exportFile(task: any, token: vscode.CancellationToken): Promise<void>;
 }