+++ src/commands/exportEntitlements.ts
@@ -1,31 +1,25 @@
 import * as vscode from 'vscode';
-import { IdentityNowResourceTreeItem, SourceTreeItem } from '../models/IdentityNowTreeItem';
+import { SourceTreeItem } from '../models/IdentityNowTreeItem';
 import { PathProposer } from '../services/PathProposer';
 import { askFile } from '../utils/vsCodeHelpers';
-import { ensureFolderExists } from '../utils/fileutils';
 import { BaseCSVExporter } from './BaseExporter';
-import { IdentityNowClient } from '../services/IdentityNowClient';
 import { Entitlement } from '../models/Entitlements';
 import EntitlementPaginator from './EntitlementPaginator';
 
 
-export class EntitlementExporter extends BaseCSVExporter<Entitlement> {
-    constructor() {
-        super("entitlements");
-    }
+export class EntitlementExporterCommand {
+
 
     /**
      * Entry point 
      * @param node 
      * @returns 
      */
-    async execute(node?: IdentityNowResourceTreeItem) {
-
+    async execute(node?: SourceTreeItem) {
         console.log("> EntitlementExporter.execute");
 
-
-        if (node === undefined || !(node instanceof SourceTreeItem)) {
-            console.log("WARNING: EntitlementExporter: invalid item", node);
+        if (node === undefined) {
+            console.error("WARNING: EntitlementExporter: invalid item", node);
             throw new Error("EntitlementExporter: invalid item");
         }
 
@@ -42,30 +36,49 @@ export class EntitlementExporter extends BaseCSVExporter<Entitlement> {
             return;
         }
 
-        ensureFolderExists(filePath);
-
-        await this.exportFileWithProgression(node.tenantName, node.tenantDisplayName, node.tenantId, filePath, node.id as string);
+        const exporter = new EntitlementExporter(
+            node.tenantId,
+            node.tenantName,
+            node.tenantDisplayName,
+            node.id as string,
+            filePath
+        );
+        await exporter.exportFileWithProgression();
     }
+}
 
-    protected async exportFile(
-        client: IdentityNowClient,
+class EntitlementExporter extends BaseCSVExporter<Entitlement> {
+
+    constructor(
+        tenantId: string,
         tenantName: string,
         tenantDisplayName: string,
         sourceId: string,
-        filePath: string,
-        task: any, token: vscode.CancellationToken): Promise<void> {
+        path: string
 
+    ) {
+        super("entitlements",
+            tenantId,
+            tenantName,
+            tenantDisplayName,
+            sourceId,
+            path);
+    }
+    
+    protected async exportFile(
+
+        task: any, token: vscode.CancellationToken): Promise<void> {
         console.log("> BaseEntitlementExporter.exportFile");
         const headers = [
-            "attributeName","attributeValue","displayName","description","privileged","schema"
+            "attributeName", "attributeValue", "displayName", "description", "privileged", "schema"
         ];
         const paths = [
-            "attribute","value","name","description","sourceSchemaObjectType"
+            "attribute", "value", "name", "description", "sourceSchemaObjectType"
         ];
         const unwindablePaths: string[] = [];
-       
-        const iterator = new EntitlementPaginator(client, sourceId);
-        await this.writeData(filePath, headers, paths, unwindablePaths, iterator);
+
+        const iterator = new EntitlementPaginator(this.client, this.sourceId);
+        await this.writeData(headers, paths, unwindablePaths, iterator, task, token);
     }
 
 }