+++ src/extension.ts
@@ -28,8 +28,8 @@ import { IdentityNowUriHandler } from './uriHandler';
 import { SortIdentityProfileCommand } from './commands/sortIdentityProfile';
 import { MenuImporter, PaletteImporter, TreeViewImporter } from './commands/importConfig';
 import { refreshIdentityProfile } from './commands/refreshIdentityProfile';
-import { AccountExporter } from './commands/exportAccounts';
-import { EntitlementExporter } from './commands/exportEntitlements';
+import { AccountExporterCommand, UncorrelatedAccountExporterCommand } from './commands/exportAccounts';
+import { EntitlementExporterCommand } from './commands/exportEntitlements';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -99,11 +99,15 @@ export function activate(context: vscode.ExtensionContext) {
 		vscode.commands.registerCommand(commands.EVALUATE_TRANSFORM,
 			(tenantTreeItem) => treeManager.evaluateTransform(tenantTreeItem)));
 
-	const accountExporterCommand = new AccountExporter();
+	const accountExporterCommand = new AccountExporterCommand();
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.EXPORT_ACCOUNTS_VIEW,
 			accountExporterCommand.execute, accountExporterCommand));
-	const entitlementExporterCommand = new EntitlementExporter();
+	const uncorrelatedAccountExporterCommand = new UncorrelatedAccountExporterCommand();
+	context.subscriptions.push(
+		vscode.commands.registerCommand(commands.EXPORT_UNCORRELATED_ACCOUNTS_VIEW,
+			uncorrelatedAccountExporterCommand.execute, uncorrelatedAccountExporterCommand));
+	const entitlementExporterCommand = new EntitlementExporterCommand();
 	context.subscriptions.push(
 		vscode.commands.registerCommand(commands.EXPORT_ENTITLEMENTS_VIEW,
 			entitlementExporterCommand.execute, entitlementExporterCommand));