+++ CHANGELOG.md
@@ -8,7 +8,11 @@ This changelog is following the recommended format by [keepachangelog](https://k
 
 ### Added 
 
-- Aggregation of entitlements
+- New command: Aggregation of entitlements
+- New command: Export of accounts
+- New command: Export of uncorrelated accounts
+- New command: Export of entitlements
+- Settings for export path
 
 ## [0.0.18] - 2023-02-15